slitter: offset each class by a different header size
Summary:
We currently rely on the base allocator, so any mismatch should
result in a glibc malloc crash or an ASan error.

This does increase the memory usage, but we'll fix that when we stop
calling into the system allocator.

DO NOT DEPLOY UNTIL THIS OVERHEAD IS REMOVED.

Depends on Dxxxx.

Checks the coronerd integration for Txxxx before adding real logic.

Test Plan: Existing CTS tests.

Reviewers: ahicks, wandrews, sbahra

Reviewed By: ahicks, wandrews

Differential Revision: https://foo.backtrace.io/Dxxxx