slitter: populate each class's cache with a "magazine"
Summary:
A magazine is a fixed-capacity array of allocation pointers, coupled
with a top-of-stack index to push/pop allocations.

For now, keep it simple, and give each cache exactly one magazine: we
don't expect our code to interleave allocation and deallocation, so we
don't need to worry about repeated magazine underflow / overflow.

Depends on Dxxxx.

Part of Txxxx.

Test Plan: New and existing tests.

Reviewers: ahicks, wandrews, sbahra

Reviewed By: ahicks

Differential Revision: https://foo.backtrace.io/Dxxxx