slitter: back object allocations with mmap
Summary:
We use a fixed alignment-based layout, where data is allocated in 2 MB
aligned chunks of 2 MB, so we can always find the chunk base.  We then
make sure to keep the chunk's metadata at a fixed offset from the
base.

Within a chunk, we satisfy individual object allocations with an
atomic bump pointer.  Chunk and their metadata are
immortal (`&'static`), so we stick the bump allocatioan fields in the
chunk, and refer to the current allocating chunk with an AtomicPtr.

Depends on Dxxxx, Dxxxx.

Part of Txxxx.

Test Plan: Existing tests. CTS tests.

Reviewers: ahicks, wandrews, sbahra

Reviewed By: ahicks

Differential Revision: https://foo.backtrace.io/Dxxxx