slitter: sanity check addresses on deallocation
Summary:
If it definitely doesn't come from the allocation class's Press,
abort noisily.

We determine that by deriving the ChunkMetadata's address from the
allocation's address, and confirming that the Class ids match.  This
may fail segfault when passed addresses that don't come from a Press,
and that's fine: that's a mismatch, and we want to crash on
mismatches.

Depends on Dxxxx.

Part of Txxxx.

Test Plan: Existing tests still pass.  Manual test crashes.

Reviewers: ahicks, wandrews, sbahra

Reviewed By: wandrews

Differential Revision: https://foo.backtrace.io/Dxxxx