slitter/mill: open the door to carving consecutive spans at once
Summary:
It's nice to have the option of small spans, but in practice, we
usually want larger spans, to amortise the span acquisition overhead.

We'll still track all the bump pointer in a single span metadata
struct, but if a single allocation corresponds to multiple spans,
we must now mark the corresponding metadata structs with their class.

Depends on Dxxxx.

Part of Txxxx.

Test Plan: Existing tests.

Reviewers: ahicks, wandrews, sbahra

Reviewed By: sbahra

Differential Revision: https://foo.backtrace.io/Dxxxx