slitter/mill: carve out runs of multiple spans at once
Summary:
The `Press` side has been ready for this: it doesn't presume to
know the number of bytes it received from the `Mill`, and knows
to initialise any "trailing" `SpanMetadata` structs.

On the Mill side, we now ask for the minimum number and ideal number
of bytes the caller would like to receive, and round that up to a
number of spans.

Depends on Dxxxx.

Finalises the heap layout and high-level architecture for Txxxx.

Test Plan:
Follow-up commit.  I want to first document all the invariants;
that should make it easier to figure out what to test / enforce
at the type level.

Reviewers: ahicks, wandrews, sbahra

Reviewed By: ahicks

Differential Revision: https://foo.backtrace.io/Dxxxx