slitter: propagate contracts more deeply
Use more `check_rep` functions to confirm that persistent structs
satisfy their internal invariants, and that their constituents are
also fine (and recursively).

This makes it possible to push more postconditions down to the
`Press`; after that, the requirements become harder to express, and
it's probably more practical to use natural language and fully manual
inspection.

Part of Txxxx.