slitter/magazine: use MaybeUninit<LinearRef> instead of Option
We really want to minimise work on the fast path, and this will
also be easier to get right when pushing / popping in C.

And add more contract checks at the same time: it's useful with
the more subtle implementation, and will be *really* helpful for
the C implementation.

Part of Txxxx.