slitter/cache: use a dedicated magazine for allocation, another for release
This new pattern means that allocate / release calls in quick
succession will not be able to pull the same object in and out of the
thread-local cache.  In return, we gain the ability to do smart stuff
about object initialisation (e.g., only zero-fill when recycling, or
only initialise to a specific bit pattern in freshly allocated
objects).

This change also lets us use a different class release mags, which in
turn makes it easier to use "empty" representations that look like
empty or full magazines, for allocation and release magazines,
respectively.

Part of Txxxx.