cache: let the magazine array be overlong
We'll want to use slices instead of a `Vec`tor to let us use a C-side
thread-local array when there aren't too many size classes.

Start by using the `per_class_info` vector as the source of truth
regarding the number of allocation classes known to the thread,
and by relaxing invariants.