slitter/c/cache: prepare for lending a slice to cache.rs
We pre-allocate a small thread-local array of `Magazines` in C,
right next to the pointer/size struct used by the fast path.  When
the pointer is actually that thread-local array, we will tend to
access addresses in the same page, if not the same cache line,
improving access locality and hopefully operation latency.